Retrieve a specific publication

A function for retrieving a specific publication. Because these are formated very differently in the API, the returning object is in a raw html_document format, best manipulated with html packages such as rvest.

Usage

get_publication(publicationid = NA, good_manners = 0)

Arguments

publicationid

Character string indicating the id of the publication to request all votes from

good_manners

Integer. Seconds delay between calls when making multiple calls to the same function

Value

A raw html_document

See also

Examples

if (FALSE) { # \dontrun{
pub <- get_publication("refs-201819-03-06")
(pub |> html_elements("replikk"))[1] |> html_text()
} # }